test(maencof): split sessionHooks spec per 3+12 rule
